Contabilita move
Applied OOP examples such as:
- Nested Classes
- Accessors
